Refactor navigation menu to iterate through destinations when creating menu.
