+++ app/src/main/java/com/example/cardboardcompanion/model/card/Card.kt
@@ -41,7 +41,7 @@ enum class CardColour(var display: String, var image: Int) {
 data class DetectedCard(
     var name: String,
     var set: String,
-    var collectorNo: Int,
+    var collectorNo: String,
     var price: Double
 ) {
