+++ app/src/main/java/com/example/cardboardcompanion/ui/component/TopNavBar.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
+import com.example.cardboardcompanion.ui.navigation.Collection
 import com.example.cardboardcompanion.ui.navigation.Destination
 import kotlinx.coroutines.launch
 
@@ -91,6 +92,9 @@ fun TopNavBar(
                         })
                     },
                     actions = {
+                        if (Collection == currentScreen){
+                            FolderSettingsMenu()
+                        }
                     }
                 )
             }
